+++ lib/src/light_stats.g.dart
@@ -1,27 +1,24 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of 'stats.dart';
+part of 'light_stats.dart';
 
 // **************************************************************************
 // JsonSerializableGenerator
 // **************************************************************************
 
-Stats _$StatsFromJson(Map<String, dynamic> json) {
-  return Stats(
+LightStats _$LightStatsFromJson(Map<String, dynamic> json) {
+  return LightStats(
     json['count'] as int,
     json['average'] as num,
-    json['median'] as num,
     json['max'] as num,
     json['min'] as num,
-    json['standardDeviation'] as num,
   );
 }
 
-Map<String, dynamic> _$StatsToJson(Stats instance) => <String, dynamic>{
+Map<String, dynamic> _$LightStatsToJson(LightStats instance) =>
+    <String, dynamic>{
       'count': instance.count,
       'average': instance.average,
-      'median': instance.median,
       'max': instance.max,
       'min': instance.min,
-      'standardDeviation': instance.standardDeviation,
     };