+++ lib/stats.dart
@@ -1,1 +1,2 @@
+export 'src/extension.dart';
 export 'src/stats.dart';